requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
twilio==8.10.0
openai==1.3.0
supabase==2.0.2
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
reportlab>=3.6.0